cron job seperated from app
