+++ .vscode/settings.json
@@ -1,5 +1,7 @@
 {
     "prettier.eslintIntegration": true,
     "prettier.printWidth": 120,
-    "prettier.tabWidth": 4
+    "prettier.tabWidth": 4,
+    "prettier.bracketSpacing": true,
+    "prettier.singleQuote": false
 }
