+++ package.json
@@ -28,14 +28,12 @@
     "lodash": "^4.17.11",
     "node-cron": "^2.0.3",
     "node-json-db": "^0.9.2",
-    "pidusage": "^2.0.17",
-    "ping": "^0.2.2",
+    "ping": "^0.2.2"
   },
   "devDependencies": {
     "eslint": "^5.15.3",
     "eslint-config-node": "^4.0.0",
     "eslint-plugin-node": "^8.0.0",
-    "jest": "^23.6.0",
-    "nodemon": "^1.18.9"
+    "jest": "^23.6.0"
   }
 }
