+++ repositories/db/index.js
@@ -1,7 +1,6 @@
 const mainDb = require("./mainDb");
 const logDb = require("./logDb");
 const settingDb = require("./settingDb");
-const processDb = require("./processDb");
 
 module.exports = {
     /**  FIXME: unitTest
@@ -36,12 +35,5 @@ module.exports = {
      */
     getLogDb(date = new Date()) {
         return logDb(date);
-    },
-
-    /** TODO: add description
-     *  
-     */
-    getProcessDb() {
-        return processDb();
     }
 };
