--- repositories/db/processDb.js
@@ -1,7 +1,0 @@
-const path = require("path");
-const JsonDB = require("node-json-db");
-const db = new JsonDB(path.join(global.__baseDirname, "database", "ProcessDb.cache.json"), true, global.isDebugMode);
-
-module.exports = () => {
-    return db;
-};
