+++ server.js
@@ -3,7 +3,6 @@ require("./utils/foldersInitialiser");
 require("./utils/classExtentions")();
 const express = require("express");
 const api = require("./api");
-const cronJobTaskRunner = require("./cron");
 const settingRepository = new (require("./repositories/settingRepository"))();
 const path = require("path");
 const cors = require("cors");
@@ -12,8 +11,8 @@ const app = express();
 app.use(cors());
 // for socket
 const http = require("http").Server(app);
-const io = require("socket.io")(http);
-const socket = require("./socket");
+// const io = require("socket.io")(http);
+// const socket = require("./socket");
 
 // manage setting for the first time
 let setting = settingRepository.getSetting();
@@ -23,7 +22,5 @@ if (!setting) setting = settingRepository.setSetting();
 app.use(express.static(path.join(__dirname, "public")));
 // start api
 api(app);
-// start socket liseners
-socket(io);
-// start cron jobs
-cronJobTaskRunner.start();
+// // start socket liseners
+// socket(io);
