+++ utils/setGlobalVariables.js
